+++ src/main/java/com/example/flabcaloriecounter/exception/FileUploadException.java
@@ -1,0 +1,8 @@
+package com.example.flabcaloriecounter.exception;
+
+public class FileUploadException extends RuntimeException {
+
+	public FileUploadException(final String message, final Throwable cause) {
+		super(message, cause);
+	}
+}
