--- src/main/java/com/example/flabcaloriecounter/feed/adapter/in/web/FeedInfoDto.java
@@ -1,8 +1,0 @@
-package com.example.flabcaloriecounter.feed.adapter.in.web;
-
-public record FeedInfoDto(
-	String contents,
-	Long userId
-) {
-
-}
