--- src/main/java/com/example/flabcaloriecounter/feed/adapter/in/web/PhotoDto.java
@@ -1,8 +1,0 @@
-package com.example.flabcaloriecounter.feed.adapter.in.web;
-
-public record PhotoDto(
-	String fileName,
-	String filePath
-) {
-
-}
