+++ src/main/java/com/example/flabcaloriecounter/feed/application/port/in/dto/PhotoDto.java
@@ -1,0 +1,8 @@
+package com.example.flabcaloriecounter.feed.application.port.in.dto;
+
+public record PhotoDto(
+	String fileName,
+	String filePath
+) {
+
+}
