--- src/main/java/com/example/flabcaloriecounter/feed/application/port/in/response/FeedDto.java
@@ -1,8 +1,0 @@
-package com.example.flabcaloriecounter.feed.application.port.in.response;
-
-public record FeedDto(
-	//  todo
-	//  String userId,
-	String contents
-) {
-}
