--- src/main/java/com/example/flabcaloriecounter/feed/application/service/ImageUploadPath.java
@@ -1,9 +1,0 @@
-package com.example.flabcaloriecounter.feed.application.service;
-
-public record ImageUploadPath(
-	//        Long postId,
-	String imageName,
-	String imagePath
-) {
-
-}
